+++ src/Stryker.Core/Stryker.Core/Initialisation/ProjectMutator.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using Stryker.Core.MutationTest;
 using Stryker.Core.Options;
 using Stryker.Core.Reporters;
@@ -26,7 +27,8 @@ namespace Stryker.Core.Initialisation
             // get a new instance of InitialisationProcess for each project
             var initialisationProcess = _initialisationProcessProvider.Provide();
             // initialize
-            var input = initialisationProcess.Initialize(options);
+            var dashboardReporter = GetDashboardReporter(reporters);
+            var input = initialisationProcess.Initialize(options, dashboardReporter);
 
             var process = _mutationTestProcessProvider.Provide(
                 mutationTestInput: input,
@@ -41,6 +43,16 @@ namespace Stryker.Core.Initialisation
             process.Mutate();
 
             return process;
+        }
+
+        private static DashboardReporter GetDashboardReporter(IReporter reporters)
+        {
+            if (reporters is BroadcastReporter broadcastReporter)
+            {
+                return broadcastReporter.Reporters.OfType<DashboardReporter>().FirstOrDefault();
+            }
+
+            return reporters as DashboardReporter;
         }
     }
 }